# Port index_select to ATen (CUDA)