--- a/aten/src/ATen/cuda/detail/IndexUtils.cuh
+++ b/aten/src/ATen/cuda/detail/IndexUtils.cuh
@@ -11,6 +11,9 @@ namespace detail {
 TORCH_CUDA_API bool maybeOverlappingIndices(const at::Tensor& t);
 TORCH_CUDA_API bool canUse32BitIndexMath(const at::Tensor &t, int64_t max_elem=std::numeric_limits<int32_t>::max());
 
+// Argument `positiveDim` turns on the legacy behavior of THC's getTensorInfo:
+// if `t` is a scalar, it's interpreted as a single-element vector.
+// (See the legacy implementation in aten/src/THC/THCTensorTypeUtils.cuh)
 template <typename scalar, typename IndexType>
 TensorInfo<scalar, IndexType>
 getTensorInfo(const at::Tensor& t, bool positiveDim = false) {