… aten/src/ATen/native/native_functions.yaml
… aten/src/THC/THCTensorIndex.cu
… aten/src/THC/generic/THCTensorIndex.cu
… test/test_torch.py