Fixes #24578