@Baranowski 2020-06-19
This is ready for a review.

ASV benchmark:

```python
import torch

class IndexSelect_1D:
    params = [None]

    def setup(self, n):
        self.src = torch.randn(10**7, device='cuda')
        self.idx = torch.rand(5 * 10**6, device='cuda').mul(10**7).floor().to(torch.long)

    def time_bench(self, n):
        self.src.index_select(0, self.idx)
        torch.cuda.synchronize('cuda')

class IndexSelect_3D:
    params = [0, 1, 2]

    def setup(self, n):
        self.src = torch.randn(500, 500, 500, device='cuda')
        self.idx = torch.rand(250, device='cuda').mul(500).floor().to(torch.long)

    def time_bench(self, n):
        self.src.index_select(n, self.idx)
        torch.cuda.synchronize('cuda')

class IndexSelect_SmallIndices:
    params = [0, 2, 4, 6]

    def setup(self, n):
        self.src = torch.randn(10, 10, 10, 10, 10, 10, 10, 10, 10, device='cuda')
        self.idx = torch.tensor([7, 3, 5, 9, 0, 1, 8, 3, 4], device='cuda')

    def time_bench(self, n):
        self.src.index_select(n, self.idx)
        torch.cuda.synchronize('cuda')
```

`master` results:

```
[ 66.67%] ··· index_select.IndexSelect_1D.time_bench                                                   ok
[ 66.67%] ··· ======== ==========
               param1
              -------- ----------
                None    1.33±0ms
              ======== ==========

[ 83.33%] ··· index_select.IndexSelect_3D.time_bench                                                   ok
[ 83.33%] ··· ======== =============
               param1
              -------- -------------
                 0      1.39±0.01ms
                 1      1.74±0.04ms
                 2      3.22±0.05ms
              ======== =============

[100.00%] ··· index_select.IndexSelect_SmallIndices.time_bench                                         ok
[100.00%] ··· ======== =============
               param1
              -------- -------------
                 0       20.1±0.7ms
                 2      19.7±0.03ms
                 4      24.4±0.03ms
                 6      34.3±0.03ms
              ======== =============
```

This PR results:

```
[ 66.67%] ··· index_select.IndexSelect_1D.time_bench                                                   ok
[ 66.67%] ··· ======== =============
               param1
              -------- -------------
                None    1.33±0.01ms
              ======== =============

[ 83.33%] ··· index_select.IndexSelect_3D.time_bench                                                   ok
[ 83.33%] ··· ======== =============
               param1
              -------- -------------
                 0      1.44±0.04ms
                 1      1.73±0.03ms
                 2      3.24±0.05ms
              ======== =============

[100.00%] ··· index_select.IndexSelect_SmallIndices.time_bench                                         ok
[100.00%] ··· ======== =============
               param1
              -------- -------------
                 0       20.2±0.5ms
                 2      19.8±0.01ms
                 4      24.3±0.05ms
                 6      34.3±0.03ms
              ======== =============
```