@Baranowski 2020-07-11
Rebased and addressed @peterbell10's comments. I don't see why the windows builds are failing. Perhaps they are flaky.