@mruberry 2020-07-14
This looks like an extremely competent port. Really cool that index_select will work for complex types on CUDA, too! (cc @anjali411) 

I have a few questions, mostly about comments and the one new arg to `getTensorInfo`. Looking forward to hearing your thoughts, @Baranowski. I should get you a response very quickly and we'll get this in!